refactor(LottoResult): 당첨금이 21억이 넘어갈 때 고려한 타입 변경

+++ src/main/java/lotto/LottoResult.java
@@ -1,16 +1,16 @@
 package lotto;
 
 public enum LottoResult {
-    NOTING(0),
-    FIFTH_PRIZE(5000),
-    FOURTH_PRIZE(50000),
-    THIRD_PRIZE(1500000),
-    SECOND_PRIZE(30000000),
-    FIRST_PRIZE(2000000000);
+    NOTING(0L),
+    FIFTH_PRIZE(5000L),
+    FOURTH_PRIZE(50000L),
+    THIRD_PRIZE(1500000L),
+    SECOND_PRIZE(30000000L),
+    FIRST_PRIZE(2000000000L);
 
-    private final int price;
+    private final long price;
 
-    LottoResult(int price) {
+    LottoResult(long price) {
         this.price = price;
     }
 }
